# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/payment_entries:
    get:
      tags:
        - PaymentEntries
      summary: List PaymentEntries
      description: Get a paginated list of PaymentEntries with optional filtering and sorting
      operationId: listPaymentEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentEntries
      summary: Create PaymentEntry
      description: Create a new PaymentEntry
      operationId: createPaymentEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PaymentEntry:
      type: object
      required:
        - id
        - payment_number
        - company_id
        - payment_type
        - posting_date
        - currency
        - method
        - paid_amount
        - allocated_amount
        - unallocated_amount
        - bank_account_id
        - party_account_id
        - withholding_amount
        - withholding_tax_type
        - status
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        payment_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        payment_type:
          type: string
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        mode_of_payment_id:
          nullable: true
          type: string
          format: uuid
        method:
          type: string
        provider_txn_id:
          nullable: true
          type: string
          format: uuid
        paid_amount:
          type: string
        allocated_amount:
          type: string
        unallocated_amount:
          type: string
        bank_account_id:
          type: string
          format: uuid
        party_account_id:
          type: string
          format: uuid
        withholding_amount:
          type: string
        withholding_account_id:
          nullable: true
          type: string
          format: uuid
        withholding_tax_type:
          type: string
        status:
          type: string
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePaymentEntryRequest:
      type: object
      required:
        - payment_number
        - company_id
        - payment_type
        - posting_date
        - currency
        - method
        - paid_amount
        - allocated_amount
        - unallocated_amount
        - bank_account_id
        - party_account_id
        - withholding_amount
        - withholding_tax_type
        - status
        - posting_state
        - metadata
      properties:
        payment_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        payment_type:
          type: string
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        mode_of_payment_id:
          nullable: true
          type: string
          format: uuid
        method:
          type: string
        provider_txn_id:
          nullable: true
          type: string
          format: uuid
        paid_amount:
          type: string
        allocated_amount:
          type: string
        unallocated_amount:
          type: string
        bank_account_id:
          type: string
          format: uuid
        party_account_id:
          type: string
          format: uuid
        withholding_amount:
          type: string
        withholding_account_id:
          nullable: true
          type: string
          format: uuid
        withholding_tax_type:
          type: string
        status:
          type: string
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PaymentEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []